--- test-data/src/main/resources/org/openmrs/module/fhir2/api/dao/impl/FhirDiagnosticReport_2_2_intial_data.xml
@@ -1,21 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<!--
-    This Source Code Form is subject to the terms of the Mozilla Public License,
-    v. 2.0. If a copy of the MPL was not distributed with this file, You can
-    obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
-    the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
-
-    Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
-    graphic logo is a trademark of OpenMRS Inc.
--->
-<dataset>
-	<concept_reference_source concept_source_id="6" name="DiagnosticServiceSectionCodes" description="DiagnosticServiceSectionCodes" creator="1" date_created="2005-01-01 00:00:00.0" retired="0" uuid="2b3c1ff8-768a-102f-83f4-12313b04a615" />
-	<fhir_concept_source fhir_concept_source_id="1" name="CIEL" url="https://openconceptlab.org/orgs/CIEL/sources/CIEL" concept_source_id="6" creator="1" date_created="2005-01-01 00:00:00.0" retired="0" uuid="0d30bea3-4ba2-4ab4-ac7b-5525840bde20" />
-	<concept concept_id="5085" retired="false" datatype_id="1" class_id="1" is_set="false" creator="1" date_created="2004-08-12 00:00:00.0" version="" changed_by="1" date_changed="2005-02-25 11:43:43.0" uuid="5085AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"/>
-	<concept_description concept_description_id="37" concept_id="5085" description="A patient's systolic blood pressure measurement (taken with a manual cuff in either a sitting or standing position)" locale="en_GB" creator="1" date_created="2004-08-12 00:00:00.0" uuid="0fa1e924-f3a0-4061-a5a5-a4cc70cde0a3"/>
-	<concept_reference_term concept_reference_term_id="12" concept_source_id="6" code="5085" name="" description="" retired="0" creator="1" date_created="2004-08-12 00:00:00.0" uuid="23d2e1cd-58d2-4972-bc86-fb21f6f81c06"/>
-	<concept_reference_map concept_map_id="11" concept_id="5085" concept_reference_term_id="12" concept_map_type_id="2" creator="1" date_created="2004-08-12 00:00:00.0" uuid="221038c1-8c9f-481a-93c0-0255ac765e48"/>
-	<concept_name concept_id="5085" name="FHIR Diagnostic Service Section Codes" locale="en_GB" creator="1" date_created="2004-08-12 00:00:00.0" concept_name_id="1439" concept_name_type="FULLY_SPECIFIED" locale_preferred="1" voided="false" uuid="67340528-feea-4232-b98c-b443c6affb39"/>
-	<fhir_diagnostic_report diagnostic_report_id="1" concept_id="5085" status="PRELIMINARY" subject_id="7" encounter_id="3" issued="2008-07-01 00:00:00.0" creator="1" date_created="2008-08-18 14:09:35.0" voided="false" uuid="1e589127-f391-4d0c-8e98-e0a158b2be22"/>
-	<fhir_diagnostic_report diagnostic_report_id="2" concept_id="5085" status="FINAL" subject_id="7" encounter_id="3" issued="2008-07-01 00:00:00.0" creator="1" date_created="2008-08-18 14:09:35.0" changed_by="1" date_changed="2018-08-18 14:09:35.0" voided="false" uuid="cf9b1f44-0e8f-42f1-900b-bf1c5d4ed590"/>
-</dataset>
